Add 'gidAllocate' storageClass param for disabling gid allocation

+++ aws/efs/cmd/efs-provisioner/efs-provisioner.go
@@ -124,9 +124,29 @@ func (p *efsProvisioner) Provision(options controller.VolumeOptions) (*v1.Persis
 		return nil, fmt.Errorf("claim.Spec.Selector is not supported")
 	}
 
-	gid, err := p.allocator.AllocateNext(options)
-	if err != nil {
-		return nil, err
+	gidAllocate := true
+	var err error
+	for k, v := range options.Parameters {
+		switch strings.ToLower(k) {
+		case "gidmin":
+		// Let allocator handle
+		case "gidmax":
+		// Let allocator handle
+		case "gidallocate":
+			gidAllocate, err = strconv.ParseBool(v)
+			if err != nil {
+				return nil, fmt.Errorf("invalid value %s for parameter %s: %v", v, k, err)
+			}
+		}
+	}
+
+	var gid *int
+	if gidAllocate {
+		allocate, err := p.allocator.AllocateNext(options)
+		if err != nil {
+			return nil, err
+		}
+		gid = &allocate
 	}
 
 	err = p.createVolume(p.getLocalPath(options), gid)
@@ -137,9 +157,6 @@ func (p *efsProvisioner) Provision(options controller.VolumeOptions) (*v1.Persis
 	pv := &v1.PersistentVolume{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: options.PVName,
-			Annotations: map[string]string{
-				gidallocator.VolumeGidAnnotationKey: strconv.FormatInt(int64(gid), 10),
-			},
 		},
 		Spec: v1.PersistentVolumeSpec{
 			PersistentVolumeReclaimPolicy: options.PersistentVolumeReclaimPolicy,
@@ -156,12 +173,20 @@ func (p *efsProvisioner) Provision(options controller.VolumeOptions) (*v1.Persis
 			},
 		},
 	}
+	if gidAllocate {
+		pv.ObjectMeta.Annotations = map[string]string{
+			gidallocator.VolumeGidAnnotationKey: strconv.FormatInt(int64(*gid), 10),
+		}
+	}
 
 	return pv, nil
 }
 
-func (p *efsProvisioner) createVolume(path string, gid int) error {
-	perm := os.FileMode(0771 | os.ModeSetgid)
+func (p *efsProvisioner) createVolume(path string, gid *int) error {
+	perm := os.FileMode(0777)
+	if gid != nil {
+		perm = os.FileMode(0771 | os.ModeSetgid)
+	}
 
 	if err := os.MkdirAll(path, perm); err != nil {
 		return err
@@ -173,11 +198,13 @@ func (p *efsProvisioner) createVolume(path string, gid int) error {
 		return err
 	}
 
-	cmd := exec.Command("chgrp", strconv.Itoa(gid), path)
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		os.RemoveAll(path)
-		return fmt.Errorf("chgrp failed with error: %v, output: %s", err, out)
+	if gid != nil {
+		cmd := exec.Command("chgrp", strconv.Itoa(*gid), path)
+		out, err := cmd.CombinedOutput()
+		if err != nil {
+			os.RemoveAll(path)
+			return fmt.Errorf("chgrp failed with error: %v, output: %s", err, out)
+		}
 	}
 
 	return nil
